requirements.txt
keyboard==0.13.5
mss==10.0.0
numpy==2.2.3
Pillow==11.1.0
pyautogui==0.9.54
